project_New

+++ public/Inc/key.h
@@ -10,10 +10,10 @@
 
 typedef struct{
 	
-	_Bool     	gpio ; 
-	uint8_t 	state ; 
-	char 		kind ; 
-	uint8_t     count  ; 
+	_Bool gpio ; 
+	uint8_t state ; 
+	char kind ; 
+	uint16_t count ; 
 
 }KEY ;
